OV-126: + update ProtectedRoute to redirect to sign-in if no token exists
